refactor: ♻️ extract spawn_session_monitor to monitor.rs (#191)
